requirements.txt
experta==1.9.4
frozendict==1.2
numpy==1.20.0
pandas==1.2.1
Pillow==8.1.0
pymongo==3.11.2
python-dateutil==2.8.1
pytz==2021.1
schema==0.6.7
six==1.15.0
